telbot/requirements.txt
APScheduler==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
boto3==1.20.31
botocore==1.23.31
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
jmespath==0.10.0
mysql-connector-python==8.0.27
protobuf==3.19.1
pyaes==1.6.1
pyasn1==0.4.8
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
rsa==4.8
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
Telethon==1.24.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
